cAPI/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
axios==0.4.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.8
datasets==2.13.0
dill==0.3.6
distlib==0.3.6
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
dnspython==2.3.0
docker==6.1.3
email-validator==2.0.0.post2
et-xmlfile==1.1.0
fastapi==0.98.0
ffmpy==0.3.0
filelock==3.12.2
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gradio==3.35.2
gradio_client==0.2.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jwt==1.3.1
kiwisolver==1.4.4
langchain==0.0.247
langchainplus-sdk==0.0.10
linkify-it-py==2.0.2
llama-index==0.6.13
loguru==0.7.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
motor==3.2.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
netcat==0.0.1a0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.1
packaging==23.1
pandas==2.0.2
pandas-stubs==2.0.2.230605
passlib==1.7.4
Pillow==9.5.0
pinecone-client==2.2.2
pipenv==2023.6.18
platformdirs==3.5.3
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.0.1
pymongo==4.4.0
pyparsing==3.0.9
pypdf==3.9.0
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-multipart==0.0.6
pytz==2022.6
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
safetensors==0.3.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toolz==0.12.0
tqdm==4.65.0
transformers==4.30.2
types-pytz==2023.3.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.6
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
websocket-client==1.6.0
websockets==11.0.3
Werkzeug==2.3.6
win32-setctime==1.1.0
xxhash==3.2.0
yarl==1.9.2
rsa==4.9
